Additions.

+++ System.Extensions/System/Collections/Generic/WithIndexItem.cs
@@ -1,6 +1,6 @@
 ﻿namespace System.Collections.Generic {
     public static class WithIndexItem {
-        public static WithIndexItem<T> Create<T>(long Index, T Item) {
+        public static WithIndexItem<T> Create<T>(int Index, T Item) {
             return new WithIndexItem<T>(Index, Item);
         }
     }
