+++ System.Extensions/System/Collections/Generic/WithIndexItem1.cs
@@ -2,15 +2,15 @@
 
 namespace System.Collections.Generic {
     public record WithIndexItem<T> : DisplayRecord {
-        public long Index { get; init; }
+        public int Index { get; init; }
         public T Item { get; init; }
 
-        public WithIndexItem(long Index, T Item) {
+        public WithIndexItem(int Index, T Item) {
             this.Index = Index;
             this.Item = Item;
         }
 
-        public void Deconstruct(out long Index, out T Item) {
+        public void Deconstruct(out int Index, out T Item) {
             Index = this.Index;
             Item = this.Item;
         }
